requirements.txt
fastapi==0.115.0
requests
uvicorn[standard]==0.30.6
anthropic>=0.40.0
groq>=0.9.0
python-dotenv==1.0.1
pydantic==2.8.2
python-multipart==0.0.9
pypdf==4.3.1
python-docx==1.1.2
python-pptx==0.6.23
deepgram-sdk>=3.0.0
openai-whisper
sounddevice
soundfile
numpy
pyttsx3